Reapply "actualizacion proveedor y ventas"
This reverts commit 2165a75b169e8cabf7aadf0866b0240e526085d6.

+++ Papeleria/src/main/java/co/edu/unbosque/Papeleria/controllers/VentaController.java
@@ -59,7 +59,7 @@ public class VentaController {
 	}
 	
 	@GetMapping("/list_sale/{id}")
-	public ResponseEntity<Optional> listCostumbyId(@PathVariable Long id) {
+	public ResponseEntity<Optional> listCostumbyId(@PathVariable int id) {
 		Optional<Venta> logs = saleDao.searchSale(id);
 		if (logs == null) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
